+++ app/src/main/java/com/zaomeng/zaomeng/model/repository/http/bean/GoodsListRowsBean.java
@@ -34,8 +34,8 @@ public class GoodsListRowsBean {
     private double standPrice;
     private String priceUnit;
     private double realPrice;
-    private double stockNumber;
-    private double stockOut;
+    private int stockNumber;
+    private int stockOut;
     private double showPrice;
     private String unitDescription;
     private String objectFeatureItemID1;
@@ -120,19 +120,19 @@ public class GoodsListRowsBean {
         this.realPrice = realPrice;
     }
 
-    public double getStockNumber() {
+    public int getStockNumber() {
         return stockNumber;
     }
 
-    public void setStockNumber(double stockNumber) {
+    public void setStockNumber(int stockNumber) {
         this.stockNumber = stockNumber;
     }
 
-    public double getStockOut() {
+    public int getStockOut() {
         return stockOut;
     }
 
-    public void setStockOut(double stockOut) {
+    public void setStockOut(int stockOut) {
         this.stockOut = stockOut;
     }
