+++ app/src/main/java/com/zaomeng/zaomeng/model/repository/http/bean/NavigatorBean.java
@@ -100,8 +100,8 @@ public class NavigatorBean {
         private String objectName;
         private String listImage;
         private double showPrice;
-        private double stockNumber;
-        private double stockOut;
+        private int stockNumber;
+        private int stockOut;
         private String unitDescription;
         private String objectFeatureItemID1;
 
@@ -137,19 +137,19 @@ public class NavigatorBean {
             this.showPrice = showPrice;
         }
 
-        public double getStockNumber() {
+        public int getStockNumber() {
             return stockNumber;
         }
 
-        public void setStockNumber(double stockNumber) {
+        public void setStockNumber(int stockNumber) {
             this.stockNumber = stockNumber;
         }
 
-        public double getStockOut() {
+        public int getStockOut() {
             return stockOut;
         }
 
-        public void setStockOut(double stockOut) {
+        public void setStockOut(int stockOut) {
             this.stockOut = stockOut;
         }
